Generates description of dot graph nodes.

Generates description of dot graph nodes.

Usage

make_dot_nodes(
  procedures_list,
  node_prefix = NULL,
  ...,
  procedures_group_bg_color = "#ecf0f1",
  procedures_group_title_font = "Helvetica"
)

Arguments

procedures_list

List of procedures. See sandardize for how to specify of procedures list.

node_prefix

pefix for node IDs

...

Arguments passed on to paste_dot_node

procedure

Procedure's function either as charater string or as call (with optional arguments)

procedure_title

This will overwrites the node's title. By default the title is either from procedure's attribute 'title' (prefixed by procedure_attr_prefix) or just a name of procedure if there is no such attibute.

id

Node ID that is used for refering in edges (arrows) definition

node_bg_color

node's background color

node_header_color

node's header background color

node_font

main font family

node_fix_width_font

fixed with font family

add_example

wheather to add example from procedure's attribure 'example' (prefixed by procedure_attr_prefix)

add_procedure_name

wheather to add procedure's name

add_procedure_args

wheather to add optional procedure's arguments

procedure_attr_prefix

prefix used to specify procedures' attibutes

procedures_group_bg_color

Background color for visuallizing groupped procedures (nested list of procedures)

procedures_group_title_font

Font family for group of procedures title

Value

dot nodes as string