Does nothing but stores (as its own default arguments) options that control vector handeling through standardization process. These options are available in most nstandr functions that accept ... parameter.

Does nothing but stores (as its own default arguments) options that control vector handeling through standardization process. These options are available in most nstandr functions that accept ... parameter.

procedure_index - replace 'procedure_index' with a number of current standardization procedure is in the list of procedures

procedure_name - replace 'procedure_name' with name of current standardization procedure

_procedure_index - replace 'procedure_index' with a number of current standardization procedure is in the list of procedures, prefixed with ''

procedure_name - replace 'procedure_name' with name of current standardization procedure, prefixed with ''

_col_name - replace 'col_name' with the name of 'col' argument (i.e., col with input value), prefixed with ''

col_name_ - replace 'col_name_' with the name of 'col' argument (i.e., col with input value), suffixed with '_'

col_name - replace 'col_name' with the name of 'col' argument (i.e., col with input value)

_index_suffix - Adds unique suffix (increment index) if the column name already exist. E.g. the following format 'col_name_index_suffix' for names(x) = c('V', 'V_9', 'V_41') and col = 2 (i.e. col_name = 'V_9' in this example) would result in a new column named 'V_42'

standardize_options(
  col = 1,
  rows = NULL,
  omitted_rows_value = NULL,
  output_placement = "replace_col",
  x_atomic_name = "x",
  output_col_name = "std{_col_name}",
  append_output_copy = FALSE,
  output_copy_col_name = "std{_col_name}_copy{_procedure_index}{_procedure_name}"
)

Arguments

col

Column of interest (the one we need to standardize) in the x object (if it is data.frame like).

rows

Logical vector to filter records of interest. Default is NULL which means do not filter records.

omitted_rows_value

If rows parameter is set then merge omitted_rows_value with the results (filtered by rows). Either single string or a character vector of length nrow(x). If NULL (the default) then original values of col are merged with results.

output_placement

Where to inset retults (standardized vector) in the x object. Default options is 'replace_col' which overwrides the col in x with results. Other options:

'omit' :: do not write results back to table (usually used when append_output_copy is set for temporary values)

'prepend_to_col' :: prepend to col

'append_to_col' :: append to col

'prepend_to_x' :: prepend to x data.frame like object

'append_to_x' :: append to x data.frame like object

x_atomic_name

If x is vector use this name for original column if it is in results. Default is "x". If x is table the name of col will be used.

output_col_name

Use this name for the column with results (standardized values). Parts in curly brakeds are substitute strings. Options for substitutions are:

append_output_copy

Whether to append a copy of result vector to x object

output_copy_col_name

How the append copy wiil be named

Value

It does nothing. Returns NULL silently.